+++ src/client/java/dev/whippet/whippets/client/BullyEntityRenderState.java
@@ -1,0 +1,8 @@
+package dev.whippet.whippets.client;
+
+import net.minecraft.client.render.entity.state.LivingEntityRenderState;
+
+public class BullyEntityRenderState extends LivingEntityRenderState {
+	/** How far through a bite it is, nought to one. */
+	public float lungeProgress;
+}