+++ src/client/java/dev/whippet/whippets/client/WhippetModelLayers.java
@@ -9,6 +9,8 @@ public final class WhippetModelLayers {
 	public static final EntityModelLayer WHIPPET_BABY = new EntityModelLayer(Whippets.id("whippet_baby"), "main");
 	public static final EntityModelLayer WHIPPET_BALL = new EntityModelLayer(Whippets.id("whippet_ball"), "main");
 
+	public static final EntityModelLayer BULLY = new EntityModelLayer(Whippets.id("bully"), "main");
+
 	public static final EntityModelLayer SQUIRREL = new EntityModelLayer(Whippets.id("squirrel"), "main");
 	public static final EntityModelLayer SQUIRREL_BABY = new EntityModelLayer(Whippets.id("squirrel_baby"), "main");